refactor: extract ScoreRange component and rename NetPromoter to ExpandedModal
